+++ bridge-app/build-app.sh
@@ -76,7 +76,7 @@ cat > "${APP_DIR}/Contents/Info.plist" << PLIST
   <key>CFBundleVersion</key>
     <string>3.5</string>
   <key>CFBundleShortVersionString</key>
-    <string>3.7</string>
+    <string>3.8</string>
   <key>PluginVersion</key>
     <string>${PLUGIN_VERSION}</string>
   <key>CFBundlePackageType</key>